+++ src/Color.cpp
@@ -3,21 +3,21 @@
 #include <random>
 #include <stdexcept>
 namespace RenderEngine {
-double r, g, b;
+float r, g, b;
 
-[[nodiscard]] double Color::clampChannel(double channel) noexcept {
-  return channel > 1. ? 1. : channel;
+[[nodiscard]] float Color::clampChannel(float channel) noexcept {
+  return channel > 1.f ? 1.f : channel;
 }
 
- Color::Color(double red, double green, double blue)
+ Color::Color(float red, float green, float blue)
     : r(red), g(green), b(blue) {
-  if (red < 0. || green < 0. || blue < 0.)
+  if (red < 0.f || green < 0.f || blue < 0.f)
     throw std::runtime_error("Color values cannot be negative");
 }
 
- Color::Color(double grayValue)
+ Color::Color(float grayValue)
     : r(grayValue), g(grayValue), b(grayValue) {
-  if (grayValue < 0.)
+  if (grayValue < 0.f)
     throw std::runtime_error("Color values cannot be negative");
 }
 
@@ -28,17 +28,17 @@ double r, g, b;
   return ColorView<u16>(*this);
 }
 
-[[nodiscard]] Color operator*(const Color &c, double val) {
-  if (val < 0.)
+[[nodiscard]] Color operator*(const Color &c, float val) {
+  if (val < 0.f)
     throw std::runtime_error("Color values cannot be negative");
   return Color{c.red() * val, c.green() * val, c.blue() * val};
 }
-[[nodiscard]] Color operator*(double val, const Color &c) {
+[[nodiscard]] Color operator*(float val, const Color &c) {
   return c * val;
 }
 
- Color &Color::operator*=(double val) {
-  if (val < 0.)
+ Color &Color::operator*=(float val) {
+  if (val < 0.f)
     throw std::runtime_error("Color values cannot be negative");
   r *= val;
   g *= val;
@@ -46,13 +46,13 @@ double r, g, b;
   return *this;
 }
 
-[[nodiscard]] double Color::red() const noexcept { return r; }
-[[nodiscard]] double Color::green() const noexcept { return g; }
-[[nodiscard]] double Color::blue() const noexcept { return b; }
+[[nodiscard]] float Color::red() const noexcept { return r; }
+[[nodiscard]] float Color::green() const noexcept { return g; }
+[[nodiscard]] float Color::blue() const noexcept { return b; }
 
-[[nodiscard]] double &Color::red() noexcept { return r; }
-[[nodiscard]] double &Color::green() noexcept { return g; }
-[[nodiscard]] double &Color::blue() noexcept { return b; }
+[[nodiscard]] float &Color::red() noexcept { return r; }
+[[nodiscard]] float &Color::green() noexcept { return g; }
+[[nodiscard]] float &Color::blue() noexcept { return b; }
 
 [[nodiscard]] Color
 Color::elementWiseMultiplication(const Color &lhs, const Color &rhs) {
@@ -66,37 +66,37 @@ Color::elementWiseMultiplication(const Color &lhs, const Color &rhs) {
 [[nodiscard]] inline Color randomColor() {
   static std::random_device rd;
   static std::mt19937 gen(rd());
-  static std::uniform_real_distribution<double> realDist(0., 1.);
+  static std::uniform_real_distribution<float> realDist(0.f, 1.f);
   static std::uniform_int_distribution<u8> intDist(0, 5);
 
   u8 variation = intDist(gen);
-  double offset = realDist(gen);
+  float offset = realDist(gen);
 
   switch (variation) {
   case 0:
-    return Color{1., offset, 0.};
+    return Color{1.f, offset, 0.f};
   case 1:
-    return Color{1. - offset, 1., 0.};
+    return Color{1.f - offset, 1.f, 0.f};
   case 2:
-    return Color{0., 1., offset};
+    return Color{0.f, 1.f, offset};
   case 3:
-    return Color{0., 1. - offset, 1.};
+    return Color{0.f, 1.f - offset, 1.f};
   case 4:
-    return Color{offset, 0., 1.};
+    return Color{offset, 0.f, 1.f};
   default:
-    return Color{1., 0, 1. - offset};
+    return Color{1.f, 0, 1.f - offset};
   }
 }
 
 } // namespace RenderEngine
 
 namespace RenderEngine::Colors {
-Color Black = Color(0., 0., 0.);
-Color Red = Color(1., 0., 0.);
-Color Green = Color(0., 1., 0.);
-Color Blue = Color(0., 0., 1.);
-Color Yellow = Color(1., 1., 0.);
-Color Cyan = Color(0., 1., 1.);
-Color Purple = Color(1., 0., 1.);
-Color White = Color(1., 1., 1.);
+Color Black = Color(0.f, 0.f, 0.f);
+Color Red = Color(1.f, 0.f, 0.f);
+Color Green = Color(0.f, 1.f, 0.f);
+Color Blue = Color(0.f, 0.f, 1.f);
+Color Yellow = Color(1.f, 1.f, 0.f);
+Color Cyan = Color(0.f, 1.f, 1.f);
+Color Purple = Color(1.f, 0.f, 1.f);
+Color White = Color(1.f, 1.f, 1.f);
 } // namespace RenderEngine::Colors
